+++ libs/rxjs/utility/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- `fromEventSource` Observable that emits values from a
+  passed [EventSource](https://developer.mozilla.org/en-US/docs/Web/API/EventSource)
+
 ## [4.3.1] - 2021-01-25
 
 ### Fixed